# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Weather Radar API",
    "version": "5.0.0"
  },
  "paths": {
    "/api/weather/current": {
      "get": {
        "tags": [
          "weather"
        ],
        "summary": "Current",
        "operationId": "current_api_weather_current_get",
        "parameters": [
          {
            "name": "stationId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Stationid"
            }
          },
          {
            "name": "units",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "e",
                "m"
              ],
              "type": "string",
              "default": "e",
              "title": "Units"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentObservationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/weather/forecast": {
      "get": {
        "tags": [
          "weather"
        ],
        "summary": "Forecast",
        "operationId": "forecast_api_weather_forecast_get",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "title": "Lat"
            }
          },
          {
            "name": "lon",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "title": "Lon"
            }
          },
          {
            "name": "units",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "e",
                "m"
              ],
              "type": "string",
              "default": "e",
              "title": "Units"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/weather/history": {
      "get": {
        "tags": [
          "weather"
        ],
        "summary": "History",
        "operationId": "history_api_weather_history_get",
        "parameters": [
          {
            "name": "stationId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Stationid"
            }
          },
          {
            "name": "units",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "e",
                "m"
              ],
              "type": "string",
              "default": "e",
              "title": "Units"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CurrentObservation": {
        "properties": {
          "stationId": {
            "type": "string",
            "title": "Stationid"
          },
          "neighborhood": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Neighborhood"
          },
          "observationTime": {
            "type": "string",
            "title": "Observationtime"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lon": {
            "type": "number",
            "title": "Lon"
          },
          "temp": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temp"
          },
          "feelsLike": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feelslike"
          },
          "dewPoint": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dewpoint"
          },
          "humidity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Humidity"
          },
          "pressure": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pressure"
          },
          "precipRate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preciprate"
          },
          "precipTotal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preciptotal"
          },
          "windSpeed": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Windspeed"
          },
          "windDir": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Winddir"
          },
          "windGust": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Windgust"
          },
          "uv": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uv"
          },
          "solarRadiation": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Solarradiation"
          },
          "units": {
            "type": "string",
            "enum": [
              "e",
              "m"
            ],
            "title": "Units"
          }
        },
        "type": "object",
        "required": [
          "stationId",
          "observationTime",
          "lat",
          "lon",
          "units"
        ],
        "title": "CurrentObservation"
      },
      "CurrentObservationResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CurrentObservation"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "CurrentObservationResponse"
      },
      "ForecastDay": {
        "properties": {
          "dayOfWeek": {
            "type": "string",
            "title": "Dayofweek"
          },
          "validDate": {
            "type": "string",
            "title": "Validdate"
          },
          "dayIcon": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dayicon"
          },
          "dayPhrase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dayphrase"
          },
          "dayPrecipChance": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dayprecipchance"
          },
          "dayTemp": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Daytemp"
          },
          "dayHumidity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dayhumidity"
          },
          "dayWindSpeed": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Daywindspeed"
          },
          "dayWindDir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Daywinddir"
          },
          "nightIcon": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nighticon"
          },
          "nightPhrase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nightphrase"
          },
          "nightPrecipChance": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nightprecipchance"
          },
          "nightTemp": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nighttemp"
          },
          "nightHumidity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nighthumidity"
          },
          "nightWindSpeed": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nightwindspeed"
          },
          "nightWindDir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nightwinddir"
          },
          "tempHigh": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temphigh"
          },
          "tempLow": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Templow"
          },
          "sunrise": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sunrise"
          },
          "sunset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sunset"
          },
          "moonPhase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Moonphase"
          }
        },
        "type": "object",
        "required": [
          "dayOfWeek",
          "validDate"
        ],
        "title": "ForecastDay"
      },
      "ForecastResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ForecastDay"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ForecastResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HistoryDay": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "tempHigh": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temphigh"
          },
          "tempLow": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Templow"
          },
          "tempAvg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tempavg"
          },
          "precipTotal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preciptotal"
          },
          "humidityHigh": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Humidityhigh"
          },
          "humidityLow": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Humiditylow"
          },
          "windSpeedAvg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Windspeedavg"
          },
          "windSpeedHigh": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Windspeedhigh"
          },
          "pressureMax": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pressuremax"
          },
          "pressureMin": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pressuremin"
          }
        },
        "type": "object",
        "required": [
          "date"
        ],
        "title": "HistoryDay"
      },
      "HistoryResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/HistoryDay"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "HistoryResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}